server/requirements.txt
openenv-core[core]>=0.2.1
torch>=2.0.0
torch-geometric>=2.3.0
sentence-transformers>=2.2.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
groq>=0.4.0
numpy>=1.24.0
